pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tokenwaster"
version = "0.1.0"
description = "A totally useless personal agent framework that wastes your tokens."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "google-genai>=0.2.0",
    "anthropic>=0.30.0",
    "pyyaml>=6.0.1",
    "rich>=13.0.0",
    "tiktoken>=0.7.0",
    "Pillow>=10.0.0",
]

[project.scripts]
tokenwaster = "tokenwaster.__main__:main"
